Encoding header negotiation deduplication and CharSequences cleanup
