+++ servicetalk-http-netty/build.gradle
@@ -44,6 +44,7 @@ dependencies {
   testImplementation testFixtures(project(":servicetalk-transport-netty-internal"))
   testImplementation project(":servicetalk-data-jackson")
   testImplementation project(":servicetalk-test-resources")
+  testImplementation project(":servicetalk-http-internal")
   testImplementation project(":servicetalk-utils-internal")
   testImplementation project(":servicetalk-concurrent-api-test")
   testImplementation project(":servicetalk-concurrent-test-internal")
