+++ settings.gradle
@@ -53,6 +53,7 @@ include "servicetalk-annotations",
         "servicetalk-grpc-protobuf",
         "servicetalk-grpc-protoc",
         "servicetalk-http-api",
+        "servicetalk-http-internal",
         "servicetalk-http-netty",
         "servicetalk-http-router-jersey",
         "servicetalk-http-router-jersey-internal",
